+++ src/components/index.js
@@ -19,6 +19,7 @@ import Pane from "./Pane";
 import Popover from "./Popover";
 import Radio from "./Radio";
 import Select from "./Select";
+import Slider from "./Slider";
 import Spinner from "./Spinner";
 import Switch from "./Switch";
 import Tab from "./Tab";
@@ -28,6 +29,7 @@ import Textarea from "./Textarea";
 import TimePicker from "./TimePicker";
 import Toastr from "./Toastr";
 import Tooltip from "./Tooltip";
+import TreeSelect from "./TreeSelect";
 import Typography from "./Typography";
 
 import "../styles/index.scss";
@@ -63,5 +65,7 @@ export {
   TimePicker,
   Toastr,
   Tooltip,
+  TreeSelect,
   Typography,
+  Slider,
 };